+++ CHANGES.md
@@ -2,10 +2,14 @@ Markdown Writer FX Change Log
 =============================
 
 ## 0.5
+
 - Move selected lines up or down with `Alt+Up` or `Alt+Down`.
 - Duplicate selected lines up or down with `Ctrl+Alt+Up` or `Ctrl+Alt+Down`.
+- Smart "Insert Header X": Support toggle header and change header level.
+
 
 ## 0.4
+
 - Find and Replace added.
 - Show line numbers in editor (if enabled in Options dialog).
 - "Save As" command added to menu bar.
@@ -13,7 +17,9 @@ Markdown Writer FX Change Log
 - Smarter "Insert Bold/Italic/Strikethrough": if used within inline code
   block, then backticks will be changed to `<code>` and `</code>`.
 
+
 ## 0.3
+
 - Replaced Markdown processor [pegdown] with [commonmark-java]
   and [flexmark-java], which are much faster and implement the [CommonMark] specification.
   [flexmark-java] is always used in the editor for syntax highlighting. 
@@ -31,14 +37,19 @@ Markdown Writer FX Change Log
 - Support \*.svg in image chooser dialog.
 - RichTextFX updated to version 0.7-M2.
 
+
 ## 0.2
+
 - RichTextFX (and dependencies) updated to version 0.6.10 (fixes bugs and memory leaks).
 - pegdown Markdown parser updated to version 1.6.
 - Added five new pegdown 1.6 extension flags to Markdown Options tab.
 - Minor improvements.
 
+
 ## 0.1
+
 - Initial release
+
 
 [CommonMark]: http://commonmark.org/
 [commonmark-java]: https://github.com/atlassian/commonmark-java